Moved logic in use weather
